[PATCH] sk98lin procfs fix
From: Christoph Hellwig <hch@infradead.org>

sk98line tries to register a procfile with the interfacename of the struct
net_device.  The patch below (ontop of the previous one) makes it work
unless you change the interface name manually, but as Linux explicitly
allows that the interface is fundamentally broken and probably should just
go away.

Signed-off-by: Andrew Morton <akpm@osdl.org>