[PATCH] 8139too: Rx fifo/overflow recovery
From: Francois Romieu <romieu@fr.zoreil.com>

This patch allows to update the interrupt status register after an Rx
overflow or a Rx fifo error even when the Rx buffer contains no packet.
The update must be kept in the packet processing loop to prevent an Rx
error storm.  As an interesting behavior, the status of the interrupt
status register must not be read early.

Signed-off-by: Francois Romieu <romieu@fr.zoreil.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>