[PATCH] USB: "Lost sync on frames" error in konicawc module
From: <fgalea@prism.uvsq.fr>

http://bugme.osdl.org/show_bug.cgi?id=3286

The kernel keeps printing "Lost sync on frames" error messages as soon as a
program tries to access the webcam.  No video data can be retrieved from
the webcam.

The following patch seems enough to solve the problem.  (just inverting the
order at which the old and new data blocks are sent to the user).

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>