[PATCH] ppc32: Switch arch/ppc/boot to lib/zlib_inflate
The following patch switches arch/ppc/boot over from using its own version
of zlib to the code found under lib/zlib_inflate.  In conjunction with the
previous two patches, the size of the resulting images isn't noticably
different.  But this does have the advantage of removing another copy of
zlib from the kernel, and I believe this allows for lib/inflate.c to go
away (as that's basically what ppc used to use).

Signed-off-by: Tom Rini <trini@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>