[PATCH] USB: khubd looks at ports after probe
Fix the hub probe logic so that when khubd wakes up, it will actually
look at all ports on the new hub.  This resolved a root hub enumeration
issue, and I expect the "boot from USB disk" folk will like this too.

The fix includes cleanup: centralizing the logic to make khubd look
at a given hub, instead of cloning it three times (or in the case of
hub probing, frankensteining it).

This also adds a FIXME to the new routine centralizing disconnect
processing:  we also want "power down port but don't wake khubd"
(for PM_SUSPEND_DISK as well as SRP).

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>