[IA64] iosapic.h, pci.c, iosapic.c, acpi.c: iosapic NUMA interrupt locality
This patch associates IOSAPICs with NUMA nodes such that interrupts
gets assigned to a reasonably good default CPU.  The patch does not
depend on the pxm_to_nid_map fixup, but results will be strange in some
configurations without it.  This should work on any NUMA box that
exposes IOSAPICs with _MAT & _PXM methods, but it's only been tested on
an rx8620.  There should be no change in behavior for boxes that don't
export both of these in ACPI namespace.
 
Signed-off-by: Alex Williamson <alex.williamson@hp.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>